Move ToView into FromIterable
Currently it is possible to convert FromIterable to collections other than View
